fix: match log topic with reuqired length
