+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules/
 lib/
-.env
+*.env
 packages/api-server/allowed-addresses.json
 crates/**/target
 target
